refactor: move test files with pipes
