+++ libs/frontend/modules/builder/src/renderer/pipes/basePropsPipe.spec.ts
@@ -1,8 +1,6 @@
-import { basePropsPipe } from '../basePropsPipe'
-import { RenderContext } from '../types'
-import { elementToRender } from './renderData'
-import { EndPipeOutput } from './types'
-import { endPipe } from './utils'
+import { basePropsPipe } from './basePropsPipe'
+import { elementToRender, endPipe, EndPipeOutput } from './test'
+import { RenderContext } from './types'
 
 const defaultContext = {} as RenderContext
 const initialProps = {}
