+++ libs/frontend/modules/builder/src/renderer/pipes/loopingRenderPipe.spec.tsx
@@ -1,9 +1,7 @@
 import { ReactElement } from 'react'
-import { loopingRenderPipe } from '../loopRenderingPipe'
-import { RenderContext } from '../types'
-import { elementToRender } from './renderData'
-import { EndPipeOutput } from './types'
-import { endPipe } from './utils'
+import { loopingRenderPipe } from './loopRenderingPipe'
+import { elementToRender, endPipe, EndPipeOutput } from './test'
+import { RenderContext } from './types'
 
 const defaultContext = {} as RenderContext
