+++ libs/frontend/modules/builder/src/renderer/pipes/propModifiersPipeline.spec.ts
@@ -1,12 +1,10 @@
 import { IHook } from '@codelab/shared/abstract/core'
 import { merge } from 'lodash'
 import { ReactElement } from 'react'
-import { RenderProps } from '../../../store'
-import { propModifiersPipeline } from '../renderPipeline'
-import { RenderContext } from '../types'
-import { elementToRender } from './renderData'
-import { EndPipeOutput } from './types'
-import { endPipe } from './utils'
+import { RenderProps } from '../../store'
+import { propModifiersPipeline } from './renderPipeline'
+import { elementToRender, endPipe, EndPipeOutput } from './test'
+import { RenderContext } from './types'
 
 const testHookResponse: RenderProps = {
   testHookResponse: {
