+++ libs/frontend/modules/builder/src/renderer/pipes/propTransformationJsPipe.spec.ts
@@ -1,9 +1,7 @@
-import { RenderProps } from '../../../store'
-import { propTransformationJsPipe } from '../propTransformationJsPipe'
-import { RenderContext } from '../types'
-import { elementToRender } from './renderData'
-import { EndPipeOutput } from './types'
-import { endPipe } from './utils'
+import { RenderProps } from '../../store'
+import { propTransformationJsPipe } from './propTransformationJsPipe'
+import { elementToRender, endPipe, EndPipeOutput } from './test'
+import { RenderContext } from './types'
 
 const defaultContext = {} as RenderContext
