+++ libs/frontend/modules/builder/src/renderer/pipes/reactNodePipe.spec.tsx
@@ -1,13 +1,17 @@
 import { TypeKind } from '@codelab/shared/abstract/core'
 import { ElementTree } from '@codelab/shared/core'
 import { render, waitFor } from '@testing-library/react'
-import { RenderProps } from '../../../store'
-import { reactNodePipe } from '../reactNodePipe'
-import { renderPipeline } from '../renderPipeline'
-import { RenderContext } from '../types'
-import { componentToRender, elementGraph, elementToRender } from './renderData'
-import { EndPipeOutput } from './types'
-import { endPipe } from './utils'
+import { RenderProps } from '../../store'
+import { reactNodePipe } from './reactNodePipe'
+import { renderPipeline } from './renderPipeline'
+import {
+  componentToRender,
+  elementGraph,
+  elementToRender,
+  endPipe,
+  EndPipeOutput,
+} from './test'
+import { RenderContext } from './types'
 
 const tree = new ElementTree(elementGraph)
