+++ libs/frontend/modules/builder/src/renderer/pipes/test/index.ts
@@ -1,2 +1,3 @@
 export * from './renderData'
 export * from './types'
+export * from './utils'
